Breaking out classification tree classes into files
