+++ src/main/scala/io/citrine/lolo/trees/classification/ClassificationTree.scala
@@ -145,80 +145,6 @@ class ClassificationResult(
   }
 }
 
-class ClassificationTrainingNode(
-                                  trainingData: Seq[(Vector[AnyVal], Char, Double)],
-                                  split: Split,
-                                  deltaImpurity: Double,
-                                  numFeatures: Int,
-                                  remainingDepth: Int,
-                                  maxDepth: Int
-                                ) extends TrainingNode(trainingData, remainingDepth) {
 
-  assert(trainingData.size > 1, "If we are going to split, we need at least 2 training rows")
-  assert(!split.isInstanceOf[NoSplit], s"Empty split split for training data: \n${trainingData.map(_.toString() + "\n")}")
 
-  lazy val (leftTrain, rightTrain) = trainingData.partition(r => split.turnLeft(r._1))
-  assert(leftTrain.size > 0 && rightTrain.size > 0, s"Split ${split} resulted in zero size: ${trainingData.map(_._1(split.getIndex()))}")
-  lazy val leftChild = if (leftTrain.size > 1 && remainingDepth > 0 && leftTrain.exists(_._2 != leftTrain.head._2)) {
-    lazy val (leftSplit, leftDelta) = ClassificationSplitter.getBestSplit(leftTrain, numFeatures)
-    if (!leftSplit.isInstanceOf[NoSplit]) {
-      new ClassificationTrainingNode(leftTrain, leftSplit, leftDelta, numFeatures, remainingDepth - 1, maxDepth)
-    } else {
-      new ClassificationTrainingLeaf(leftTrain, maxDepth - remainingDepth)
-    }
-  } else {
-    new ClassificationTrainingLeaf(leftTrain, maxDepth - remainingDepth)
-  }
-  lazy val rightChild = if (rightTrain.size > 1 && remainingDepth > 0 && rightTrain.exists(_._2 != rightTrain.head._2)) {
-    lazy val (rightSplit, rightDelta) = ClassificationSplitter.getBestSplit(rightTrain, numFeatures)
-    if (!rightSplit.isInstanceOf[NoSplit]) {
-      new ClassificationTrainingNode(rightTrain, rightSplit, rightDelta, numFeatures, remainingDepth - 1, maxDepth)
-    } else {
-      new ClassificationTrainingLeaf(rightTrain, maxDepth - remainingDepth)
-    }
-  } else {
-    new ClassificationTrainingLeaf(rightTrain, maxDepth - remainingDepth)
-  }
 
-  /**
-    * Get the lightweight prediction node for the output tree
-    *
-    * @return lightweight prediction node
-    */
-  override def getNode(): ModelNode[PredictionResult[Char]] = new InternalModelNode(
-    split, leftChild.getNode(), rightChild.getNode()
-  )
-
-  override def getFeatureImportance(): mutable.ArraySeq[Double] = {
-    val improvement = deltaImpurity
-    var ans = leftChild.getFeatureImportance().zip(rightChild.getFeatureImportance()).map(p => p._1 + p._2)
-    ans(split.getIndex) = ans(split.getIndex) + improvement
-    ans
-  }
-}
-
-class ClassificationTrainingLeaf(
-                                  trainingData: Seq[(Vector[AnyVal], Char, Double)],
-                                  depth: Int
-                                ) extends TrainingNode(trainingData, 0) {
-
-  lazy val mode: Char = trainingData.map(_._2).groupBy(identity).mapValues(_.size).maxBy(_._2)._1
-
-  /**
-    * Get the lightweight prediction node for the output tree
-    *
-    * @return lightweight prediction node
-    */
-  override def getNode(): ModelNode[PredictionResult[Char]] = new ClassificationLeaf(mode, depth)
-
-  override def getFeatureImportance(): mutable.ArraySeq[Double] = mutable.ArraySeq.fill(trainingData.head._1.size)(0.0)
-}
-
-/**
-  * The leaves just predict the most common value
-  *
-  * @param mode most common value
-  */
-class ClassificationLeaf(mode: Char, depth: Int) extends ModelNode[PredictionResult[Char]] {
-  override def transform(input: Vector[AnyVal]): (PredictionResult[Char], TreeMeta) = (MultiResult(Seq(mode)), TreeMeta(depth))
-}
